Enable motor on startup
Motor should be enabled on startup so the controller applies current for
holding torque. Also removes the need to enable the motor before any move function.
Motor should not be disabled at any time while cnc is on. Otherwise a disabled axis will
be more susceptible to unwanted movement due to a lack of holding torque.