Added a timer component to ESP files
The timer namespace will handle move operations.
A move will be setup and started in separate functions to allow for sync move.
Step timer callback will iterate through a step vector and only send pulses if the vector value is true.
This is how acceleration will be implemented (unless I come up with something better).
An equally sized direction vector will be used to set the direction for the next step, allowing for direction changes
	within the same move operation, such as for a circular move.
The step and direction vectors will be calculated in a seperate component.

Also added a 'TEST_FUNCTION' SPI command.
This will be used to test functionality on the esp, hopefully without too much extra work each time.